+++ Package.swift
@@ -14,6 +14,10 @@ let package = Package(
             swiftSettings: [
                 .unsafeFlags(["-parse-as-library"])
             ]
+        ),
+        .testTarget(
+            name: "PestyTests",
+            dependencies: ["Pesty"]
         )
     ],
     swiftLanguageModes: [.v5]